cli/azd/extensions/microsoft.azd.extensions/internal/resources/languages/python/requirements.txt
grpcio==1.44.0  
grpcio-tools==1.44.0  
protobuf==3.19.4  